+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Fix the implementation of `pg-tables` and `pg-columns` to use the information schema instead of
   historical alternative SQL queries.
 
+- Add function `pg-add-notification-handler` to add a function to the list of handlers for
+  `NotificationResponse` messages from PostgreSQL. A handler takes two arguments, the channel and
+  the payload, which correspond to SQL-level `NOTIFY channel, 'payload'`.
+
 
 ## [0.23] - 2023-08-20
 ### New